// src/components/player/types.ts
export interface BgProps {
  id: number
  name: string
  img: string
  thumbnail: string
  artist: string
  price: number
  type: string
  when: string
}

export interface ColorProps {
  id: number
  name: string
  hex: string
}

export interface TitleProps {
  id: number
  name: string
  price: number
  type: string
  when: string
}

export interface IconProps {
  id: number
  name: string
  img: string
  thumbnail: string
  price: number
  type: string
}

export interface FrameProps {
  id: number
  name: string
  img: string
  thumbnail: string
  price: number
}